requirements.txt
biopython
pandas
PubChemPy
selenium
requests